website/requirements/local.txt
-r base.txt

coverage==7.6.10
djlint==1.36.4
django-debug-toolbar
mypy==1.14.1
mypy-extensions==1.0.0
moto==5.0.26
pre_commit==4.0.1
pytest-django==4.9.0
pytest==8.3.4
ruff==0.8.4
PyPDF2==3.0.1
